+++ gui/listfield.d
@@ -211,7 +211,7 @@ class ListField : Widget, ListInterface!(String) {
 		return control_listbox.opApply(loopFunc);
 	}
 	
-	int opApply(int delegate(ref int, ref String) loopFunc) {
+	int opApply(int delegate(ref size_t, ref String) loopFunc) {
 		return control_listbox.opApply(loopFunc);
 	}
 protected:
